+++ drizzle/0007_large_the_renegades.sql
@@ -1,0 +1,3 @@
+ALTER TABLE `clients` ADD `client_no` integer;--> statement-breakpoint
+ALTER TABLE `orders` ADD `seq` integer;--> statement-breakpoint
+ALTER TABLE `orders` ADD `order_no` text;